refactor: add exception handling for asset deletion when assignments exist

+++ app/Http/Services/AssetService.php
@@ -3,6 +3,7 @@
 namespace App\Http\Services;
 
 use App\Enums\AssetStatus;
+use App\Exceptions\ClientException;
 use App\Models\Asset;
 use App\Tables\AssetTable;
 use App\Tables\Traits\HasTable;
@@ -65,6 +66,10 @@ class AssetService extends Service
     public function delete(Asset $asset)
     {
         return DB::transaction(function () use ($asset) {
+            if ($asset->assignments()->exists()) {
+                throw new ClientException('Asset cannot be deleted because it is assigned to an assignment.');
+            }
+
             $asset->delete();
         });
     }
